+++ demo/backstop-common.js
@@ -451,6 +451,31 @@ module.exports = function backstopCommon(baseUrl) {
         viewports: [{ label: 'form-field', width: 800, height: 400 }],
       },
       {
+        label: `Forms/Select (default)`,
+        url: componentUrlFor(`select/default`),
+        viewports: [{ label: 'form-field', width: 800, height: 400 }],
+      },
+      {
+        label: `Forms/Select (optional)`,
+        url: componentUrlFor(`select/optional`),
+        viewports: [{ label: 'form-field', width: 800, height: 400 }],
+      },
+      {
+        label: `Forms/Select (error)`,
+        url: componentUrlFor(`select/error`),
+        viewports: [{ label: 'form-field', width: 800, height: 400 }],
+      },
+      {
+        label: `Forms/Select (hint)`,
+        url: componentUrlFor(`select/hint`),
+        viewports: [{ label: 'form-field', width: 800, height: 400 }],
+      },
+      {
+        label: `Forms/Select (value)`,
+        url: componentUrlFor(`select/value`),
+        viewports: [{ label: 'form-field', width: 800, height: 400 }],
+      },
+      {
         label: 'Example links',
         url: componentUrlFor('links/example_links'),
       },